+++ lib/circuitry/processors/batcher.rb
@@ -2,20 +2,13 @@ require 'circuitry/processor'
 
 module Circuitry
   module Processors
-    module Batcher
-      class << self
-        include Processor
+    class Batcher < Processor
+      def process
+        # noop
+      end
 
-        def process(&block)
-          raise ArgumentError, 'no block given' unless block_given?
-          pool << block
-        end
-
-        def flush
-          while (block = pool.shift)
-            safely_process(&block)
-          end
-        end
+      def wait
+        safely_process(&block)
       end
     end
   end