+++ spec/circuitry/concerns/async_spec.rb
@@ -3,17 +3,13 @@ require 'spec_helper'
 async_class = Class.new do
   include Circuitry::Concerns::Async
 
-  def self.default_async_strategy
-    :thread
-  end
-
   def self.async_strategies
     [:fork, :thread, :batch]
   end
-end
 
-incomplete_async_class = Class.new do
-  include Circuitry::Concerns::Async
+  def config
+    Circuitry.subscriber_config
+  end
 end
 
 RSpec.describe Circuitry::Concerns::Async, type: :model do
@@ -40,20 +36,8 @@ RSpec.describe Circuitry::Concerns::Async, type: :model do
     end
 
     describe 'with true' do
-      describe 'when the class has defined a default async strategy' do
-        it 'sets async to the default value' do
-          expect(subject.class).to receive(:default_async_strategy).at_least(:once).and_call_original
-          subject.async = true
-          expect(subject.async).to eq subject.class.default_async_strategy
-        end
-      end
-
-      describe 'when the class has not defined a default async strategy' do
-        subject { incomplete_async_class.new }
-
-        it 'raises an error' do
-          expect { subject.async = true }.to raise_error(NotImplementedError)
-        end
+      it 'sets async to the default value' do
+        expect { subject.async = true }.to change { subject.async }.to(subject.config.async_strategy)
       end
     end
 
@@ -87,39 +71,50 @@ RSpec.describe Circuitry::Concerns::Async, type: :model do
   end
 
   describe '#process_asynchronously' do
-    let(:block) { ->{ } }
+    let(:block) { ->(_) { } }
+    let(:processor) { double('Circuitry::Processor', process: nil, wait: nil, is_a?: true) }
+
+    shared_examples_for 'an asynchronous processor' do
+      before do
+        allow(Circuitry::Pool).to receive(:<<)
+      end
+
+      it 'delegates to the processor' do
+        subject.process_asynchronously(&block)
+        expect(processor).to have_received(:process)
+      end
+
+      it 'adds the processor to the pool' do
+        subject.process_asynchronously(&block)
+        expect(Circuitry::Pool).to have_received(:<<).with(processor)
+      end
+    end
 
     describe 'via forking' do
       before do
         allow(subject).to receive(:async).and_return(:fork)
+        allow(Circuitry::Processors::Forker).to receive(:new).with(any_args, &block).and_return(processor)
       end
 
-      it 'delegates to fork processor' do
-        expect(Circuitry::Processors::Forker).to receive(:process).with(no_args, &block)
-        subject.process_asynchronously(&block)
-      end
+      it_behaves_like 'an asynchronous processor'
     end
 
     describe 'via threading' do
       before do
         allow(subject).to receive(:async).and_return(:thread)
+        allow(Circuitry::Processors::Threader).to receive(:new).with(any_args, &block).and_return(processor)
       end
 
-      it 'delegates to thread processor' do
-        expect(Circuitry::Processors::Threader).to receive(:process).with(no_args, &block)
-        subject.process_asynchronously(&block)
-      end
+      it_behaves_like 'an asynchronous processor'
     end
 
     describe 'via batching' do
       before do
         allow(subject).to receive(:async).and_return(:batch)
+        allow(Circuitry::Processors::Batcher).to receive(:new).with(any_args, &block).and_return(processor)
       end
 
-      it 'delegates to batch processor' do
-        expect(Circuitry::Processors::Batcher).to receive(:process).with(no_args, &block)
-        subject.process_asynchronously(&block)
-      end
+      it_behaves_like 'an asynchronous processor'
     end
   end
 end