+++ spec/support/examples/async_processor_examples.rb
@@ -6,12 +6,12 @@ RSpec.shared_examples_for 'an asyncronous processor' do
     before do
       allow(subject).to receive(:fork) { |&block| block.call }
       allow(Process).to receive(:detach)
-      allow(Circuitry.subscriber_config).to receive(:on_async_exit).and_return(on_async_exit)
+      allow(config).to receive(:on_async_exit).and_return(on_async_exit)
     end
 
     it 'calls the proc' do
-      subject.process(&block)
-      subject.flush
+      subject.process
+      subject.wait
       expect(on_async_exit).to have_received(:call)
     end
   end